Port compiledProjects saga test to Jest
